# resources/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "COPS REST API",
        "version": "3.4.0"
    },
    "servers": [
        {
            "url": "vendor/bin/index.php/restapi",
            "description": "COPS REST API Endpoint"
        }
    ],
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-KEY"
            },
            "BasicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        },
        "parameters": {
            "dbParam": {
                "name": "db",
                "in": "query",
                "required": false,
                "schema": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        }
    },
    "paths": {
...
        "/custom/{custom}": {
            "get": {
                "summary": "Route to page handler with page=14&custom={custom}",
                "operationId": "get_page-14-custom",
                "tags": [
                    "default"
                ],
                "responses": {
                    "200": {
                        "description": "Result of page handler with page=14&custom={custom}"
                    }
                },
                "parameters": [
                    {
                        "name": "custom",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "^\\d+$"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ]
            }
        },
...
        "/custom": {
            "get": {
                "summary": "Route to custom api",
                "operationId": "get_restapi-CustomColumnType",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of custom api"
                    }
                },
                "parameters": [
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ]
            }
        },
        "/databases/{db}/{name}": {
            "get": {
                "summary": "Route to databases/{db}/{name} api - with api key",
                "operationId": "get_restapi-Database-db-name",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of databases/{db}/{name} api"
                    }
                },
                "parameters": [
                    {
                        "name": "db",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/databases/{db}": {
            "get": {
                "summary": "Route to databases/{db} api",
                "operationId": "get_restapi-Database-db",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of databases/{db} api"
                    }
                },
                "parameters": [
                    {
                        "name": "db",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "table",
                                "view"
                            ]
                        },
                        "example": "table"
                    }
                ]
            }
        },
...
        "/notes/{type}/{id}/{title}": {
            "get": {
                "summary": "Route to notes/{type}/{id}/{title} api",
                "operationId": "get_restapi-Note-type-id-title",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of notes/{type}/{id}/{title} api"
                    }
                },
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ]
            }
        },
        "/notes/{type}/{id}": {
            "get": {
                "summary": "Route to notes/{type}/{id} api",
                "operationId": "get_restapi-Note-type-id",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of notes/{type}/{id} api"
                    }
                },
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ]
            }
        },
        "/notes/{type}": {
            "get": {
                "summary": "Route to notes/{type} api",
                "operationId": "get_restapi-Note-type",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of notes/{type} api"
                    }
                },
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ]
            }
        },
        "/notes": {
            "get": {
                "summary": "Route to notes api",
                "operationId": "get_restapi-Note",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of notes api"
                    }
                },
                "parameters": [
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ]
            }
        },
        "/preferences/{key}": {
            "get": {
                "summary": "Route to preferences/{key} api",
                "operationId": "get_restapi-Preference-key",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of preferences/{key} api"
                    }
                },
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ]
            }
        },
        "/preferences": {
            "get": {
                "summary": "Route to preferences api",
                "operationId": "get_restapi-Preference",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of preferences api"
                    }
                },
                "parameters": [
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ]
            }
        },
        "/annotations/{bookId}/{id}": {
            "get": {
                "summary": "Route to annotations/{bookId}/{id} api",
                "operationId": "get_restapi-Annotation-bookId-id",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of annotations/{bookId}/{id} api"
                    }
                },
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ]
            }
        },
        "/annotations/{bookId}": {
            "get": {
                "summary": "Route to annotations/{bookId} api",
                "operationId": "get_restapi-Annotation-bookId",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of annotations/{bookId} api"
                    }
                },
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ]
            }
        },
        "/annotations": {
            "get": {
                "summary": "Route to annotations api",
                "operationId": "get_restapi-Annotation",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of annotations api"
                    }
                },
                "parameters": [
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ]
            }
        },
        "/metadata/{bookId}/{element}/{name}": {
            "get": {
                "summary": "Route to metadata/{bookId}/{element}/{name} api",
                "operationId": "get_restapi-Metadata-bookId-element-name",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of metadata/{bookId}/{element}/{name} api"
                    }
                },
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "element",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ]
            }
        },
        "/metadata/{bookId}/{element}": {
            "get": {
                "summary": "Route to metadata/{bookId}/{element} api",
                "operationId": "get_restapi-Metadata-bookId-element",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of metadata/{bookId}/{element} api"
                    }
                },
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "element",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ]
            }
        },
        "/metadata/{bookId}": {
            "get": {
                "summary": "Route to metadata/{bookId} api",
                "operationId": "get_restapi-Metadata-bookId",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of metadata/{bookId} api"
                    }
                },
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ]
            }
        },
        "/user/details": {
            "get": {
                "summary": "Route to user/details api - with basic authentication",
                "operationId": "get_restapi-User-details",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of user/details api"
                    }
                },
                "parameters": [
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ],
                "security": [
                    {
                        "BasicAuth": []
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "summary": "Route to user api - with basic authentication",
                "operationId": "get_restapi-User",
                "tags": [
                    "restapi"
                ],
                "responses": {
                    "200": {
                        "description": "Result of user api"
                    }
                },
                "parameters": [
                    {
                        "$ref": "#/components/parameters/dbParam"
                    }
                ],
                "security": [
                    {
                        "BasicAuth": []
                    }
                ]
            }
        },
...
    }
}